Handle memory underflow
