+++ Sources/Error/RuntimeError.swift
@@ -1,0 +1,11 @@
+//
+//  RuntimeError.swift
+//  
+//
+//  Created by Dominik Grodl on 13.02.2024.
+//
+
+enum RuntimeError: Error {
+    case invalidInput
+}
+
