+++ Sources/Error/ValidationError.swift
@@ -1,0 +1,12 @@
+//
+//  ValidationError.swift
+//
+//
+//  Created by Dominik Grodl on 13.02.2024.
+//
+
+enum ValidationError: Error {
+    case noProgramProvided
+    case noInputPathProvided
+    case unsuportedFileType(String)
+}
